fix(data-model): corrects circular dependencies
