+++ .eslintrc.yml
@@ -55,7 +55,7 @@ rules:
   curly: error
   implicit-arrow-linebreak: 'off'
   import/extensions: 0
-  import/no-cycle: 0
+  import/no-cycle: 2
   import/prefer-default-export: 'off'
   json-files/sort-package-json: 1
   key-spacing:
