+++ src/stories/flows/complex/data/sections.flow.ts
@@ -1,5 +1,5 @@
 /* eslint-disable sonarjs/no-duplicate-string */
-import { ISection } from '../../../../survey/ISection';
+import { ISection } from '../../../../survey/IStep';
 
 export const sections: ISection[] = [
   {
