+++ src/stories/flows/simple/data/questions.flow.ts
@@ -3,7 +3,7 @@
                 sonarjs/no-duplicate-string,
  */
 import { QUESTION_TYPE } from '../../../../lib/enums';
-import { IQuestion }     from '../../../../survey/IQuestion';
+import { IQuestion }     from '../../../../survey/IStep';
 
 /**
  * All of the questions, their answers and dependencies
