+++ src/survey/IPages.ts
@@ -1,4 +1,4 @@
-import { IPage } from './IPage';
+import { IPage } from './IStep';
 
 /**
  * Defines required pages for the survey flow
