--- src/survey/IResponse.ts
@@ -1,7 +1,0 @@
-import { IQuestion }       from './IQuestion';
-import { IQuestionAnswer } from './IQuestionAnswer';
-
-export interface IResponse {
-  answers: Partial<IQuestionAnswer>[];
-  question: Partial<IQuestion>;
-}
