+++ src/survey/index.ts
@@ -2,17 +2,12 @@ export * from './IAction';
 export * from './IForm';
 export * from './IDesignData';
 export * from './INavButton';
-export * from './IPage';
 export * from './IPageData';
 export * from './IPages';
-export * from './IQuestion';
 export * from './IQuestionAnswer';
 export * from './IQuestionData';
 export * from './IQuestionableConfig';
 export * from './IQuestionnaire';
-export * from './IResponse';
-export * from './IRequirement';
 export * from './IResult';
-export * from './ISection';
 export * from './IStep';
 export * from './IStepData';
